refator: write main layout for home page and reply page
